# db: Improve performance of external service sync failure message