--- a/migrations/frontend/1657106983_faster_failure_msg_query/down.sql
+++ b/migrations/frontend/1657106983_faster_failure_msg_query/down.sql
@@ -1,0 +1,1 @@
+DROP INDEX IF EXISTS external_service_sync_jobs_state_external_service_id;