--- a/migrations/frontend/1657106983_faster_failure_msg_query/metadata.yaml
+++ b/migrations/frontend/1657106983_faster_failure_msg_query/metadata.yaml
@@ -1,0 +1,3 @@
+name: faster_failure_msg_query
+parents: [1653472246, 1655226733, 1655843069, 1655157509, 1655454264]
+createIndexConcurrently: true