--- a/migrations/frontend/1657106983_faster_failure_msg_query/up.sql
+++ b/migrations/frontend/1657106983_faster_failure_msg_query/up.sql
@@ -1,0 +1,1 @@
+CREATE INDEX CONCURRENTLY IF NOT EXISTS external_service_sync_jobs_state_external_service_id ON external_service_sync_jobs (state, external_service_id) INCLUDE (finished_at);