--- a/migrations/frontend/1657107627_drop_duplicate_index_sync_jobs_state/down.sql
+++ b/migrations/frontend/1657107627_drop_duplicate_index_sync_jobs_state/down.sql
@@ -1,0 +1,1 @@
+CREATE INDEX IF NOT EXISTS external_service_sync_jobs_state_idx ON external_service_sync_jobs (state);