--- a/migrations/frontend/1657107627_drop_duplicate_index_sync_jobs_state/metadata.yaml
+++ b/migrations/frontend/1657107627_drop_duplicate_index_sync_jobs_state/metadata.yaml
@@ -1,0 +1,2 @@
+name: drop_duplicate_index_sync_jobs_state
+parents: [1657106983]