--- a/migrations/frontend/1657107627_drop_duplicate_index_sync_jobs_state/up.sql
+++ b/migrations/frontend/1657107627_drop_duplicate_index_sync_jobs_state/up.sql
@@ -1,0 +1,1 @@
+DROP INDEX IF EXISTS external_service_sync_jobs_state_idx;